Refactored the ILogContext/EnabledActivitySource to just use it, if present, and avoid the overhead of checking, etc. which is no longer best practice with OT.
