+++ src/MassTransit/InMemoryTransport/InMemoryTransport/InMemorySendTransport.cs
@@ -33,7 +33,7 @@ namespace MassTransit.InMemoryTransport
 
             await pipe.Send(context).ConfigureAwait(false);
 
-            StartedActivity? activity = LogContext.IfEnabled(_context.ActivityName)?.StartSendActivity(context);
+            StartedActivity? activity = LogContext.Current?.StartSendActivity(_context, context);
             try
             {
                 if (_context.SendObservers.Count > 0)
