+++ src/MassTransit/Middleware/InitiatedBySagaMessageFilter.cs
@@ -23,7 +23,7 @@ namespace MassTransit.Middleware
 
         public async Task Send(SagaConsumeContext<TSaga, TMessage> context, IPipe<SagaConsumeContext<TSaga, TMessage>> next)
         {
-            StartedActivity? activity = LogContext.IfEnabled(OperationName.Saga.Initiate)?.StartSagaActivity(context);
+            StartedActivity? activity = LogContext.Current?.StartSagaActivity(context);
             try
             {
                 await context.Saga.Consume(context).ConfigureAwait(false);
