+++ src/MassTransit/Middleware/ObservesSagaMessageFilter.cs
@@ -22,7 +22,7 @@ namespace MassTransit.Middleware
 
         public async Task Send(SagaConsumeContext<TSaga, TMessage> context, IPipe<SagaConsumeContext<TSaga, TMessage>> next)
         {
-            StartedActivity? activity = LogContext.IfEnabled(OperationName.Saga.Observe)?.StartSagaActivity(context);
+            StartedActivity? activity = LogContext.Current?.StartSagaActivity(context);
             try
             {
                 await context.Saga.Consume(context).ConfigureAwait(false);
