+++ src/MassTransit/Middleware/OrchestratesSagaMessageFilter.cs
@@ -23,7 +23,7 @@
 
         public async Task Send(SagaConsumeContext<TSaga, TMessage> context, IPipe<SagaConsumeContext<TSaga, TMessage>> next)
         {
-            StartedActivity? activity = LogContext.IfEnabled(OperationName.Saga.Orchestrate)?.StartSagaActivity(context);
+            StartedActivity? activity = LogContext.Current?.StartSagaActivity(context);
             try
             {
                 await context.Saga.Consume(context).ConfigureAwait(false);
