+++ src/MassTransit/Transports/BaseSendTransportContext.cs
@@ -13,6 +13,7 @@ namespace MassTransit.Transports
         SendTransportContext
     {
         readonly Lazy<string> _activityName;
+        readonly Lazy<string> _destination;
         readonly IHostConfiguration _hostConfiguration;
 
         protected BaseSendTransportContext(IHostConfiguration hostConfiguration, ISerialization serialization)
@@ -23,21 +24,23 @@ namespace MassTransit.Transports
 
             Serialization = serialization;
 
-            _activityName = new Lazy<string>(() =>
+            _destination = new Lazy<string>(() =>
             {
                 var endpointName = EntityName;
 
                 if (endpointName.Contains("_bus_"))
                     endpointName = "bus";
-                else if(endpointName.Contains("_endpoint_"))
+                else if (endpointName.Contains("_endpoint_"))
                     endpointName = "endpoint";
-                else if(endpointName.Contains("_signalr_"))
+                else if (endpointName.Contains("_signalr_"))
                     endpointName = "signalr";
-                else if(endpointName.StartsWith("Instance_"))
+                else if (endpointName.StartsWith("Instance_"))
                     endpointName = "instance";
 
-                return $"{endpointName} send";
+                return endpointName;
             });
+
+            _activityName = new Lazy<string>(() => $"{_destination.Value} send");
         }
 
         public abstract string EntityName { get; }
@@ -45,6 +48,7 @@ namespace MassTransit.Transports
         public ILogContext LogContext => _hostConfiguration.SendLogContext ?? throw new InvalidOperationException("SendLogContext should not be null");
 
         public string ActivityName => _activityName.Value;
+        public string ActivityDestination => _destination.Value;
 
         public SendObservable SendObservers { get; }
