+++ src/Transports/MassTransit.ActiveMqTransport/ActiveMqTransport/ActiveMqSendTransport.cs
@@ -82,7 +82,7 @@
 
                 await _pipe.Send(context).ConfigureAwait(false);
 
-                StartedActivity? activity = LogContext.IfEnabled(_context.ActivityName)?.StartSendActivity(context);
+                StartedActivity? activity = LogContext.Current?.StartSendActivity(_context, context);
                 try
                 {
                     if (_context.SendObservers.Count > 0)
