+++ src/Transports/MassTransit.Azure.ServiceBus.Core/AzureServiceBusTransport/ServiceBusSendTransport.cs
@@ -75,7 +75,7 @@
 
                 CopyIncomingIdentifiersIfPresent(context);
 
-                StartedActivity? activity = LogContext.IfEnabled(_context.ActivityName)?.StartSendActivity(context,
+                StartedActivity? activity = LogContext.Current?.StartSendActivity(_context, context,
                     (nameof(context.PartitionKey), context.PartitionKey),
                     (nameof(context.SessionId), context.SessionId));
                 try
